formated the mail

+++ src/main/java/com/smartparking/smart_parking_management_system/dto/ReservationRequest.java
@@ -7,6 +7,6 @@ public class ReservationRequest  {
     // private Long userId;
     private Long parkingSpotId;
     private String startTime;
-    private int durationInHours;
+    private double durationInHours;
     
 }
