+++ src/main/java/com/smartparking/smart_parking_management_system/service/ReservationService.java
@@ -6,7 +6,7 @@ import java.util.List;
 import com.smartparking.smart_parking_management_system.model.Reservation;
 
 public interface ReservationService {
-    Reservation createReservation(Long parkingSpotId,LocalDateTime requestedStartTime, int durationInHours) throws Exception;
+    Reservation createReservation(Long parkingSpotId,LocalDateTime requestedStartTime, double durationInHours) throws Exception;
     List<Reservation> getReservationByUser(Long userId);
     List<Reservation> getAllReservations();
